Secp256k1Constants: move to `o.b.base` module
These constants will actually be useful in `base` and perhaps in the
future could even be made part of the API. For now lets keep them
in the `o.b.base.internal` package.
