--- core/src/main/java/org/bitcoinj/crypto/secp/package-info.java
@@ -1,24 +1,0 @@
-/*
- * Copyright by the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * This is a <b>temporary</b> package for holding temporary classes and interfaces that we anticipate
- * will eventually be replaced by classes and interfaces in <b>secp256k1-jdk</b>
- */
-@NullMarked
-package org.bitcoinj.crypto.secp;
-
-import org.jspecify.annotations.NullMarked;
